+++ components/Footer.tsx
@@ -1,0 +1,46 @@
+import Image from "next/image";
+
+export default function Footer(props) {
+    return (
+        <>
+        <footer>
+            <a
+            href="https://vercel.com?utm_source=create-next-app&utm_medium=default-template&utm_campaign=create-next-app"
+            target="_blank"
+            rel="noopener noreferrer"
+            >
+            Powered by{' '}
+            <Image src="/vercel.svg" alt="Picture of the product" width={200} height={200} className="logo"/>
+            </a>
+        </footer>
+
+        <style jsx>{`
+        footer {
+          width: 100%;
+          height: 100px;
+          border-top: 1px solid #eaeaea;
+          display: flex;
+          justify-content: center;
+          align-items: center;
+        }
+
+        footer img {
+          margin-left: 0.5rem;
+        }
+
+        footer a {
+          display: flex;
+          justify-content: center;
+          align-items: center;
+        }
+        @media (max-width: 600px) {
+          .grid {
+            width: 100%;
+            flex-direction: column;
+          }
+        }
+      `}</style>
+        </>
+
+    );
+}
