+++ pages/_app.tsx
@@ -1,5 +1,5 @@
 import Head from 'next/head';
-// This default export is required in a new `pages/_app.js` file.
+// This default export is required in a new `pages/_app.tsx` file.
 export default function MyApp({ Component, pageProps }) {
     return (
         <div className="root-app">
